React Native Files
This is what was installed when I installed React Native. I also deleted the files relating to just React. I think you don't need to install React Native and go through that whole process if you fetch this, but I'm not sure. Check by running cd ShowerApp and then npm run ios or a similar command in the terminal.